adding compare beacon block test
