+++ testing/endtoend/endtoend_setup_test.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/prysmaticlabs/prysm/v3/config/params"
 	ev "github.com/prysmaticlabs/prysm/v3/testing/endtoend/evaluators"
+	"github.com/prysmaticlabs/prysm/v3/testing/endtoend/evaluators/beacon-api_v1"
 	e2eParams "github.com/prysmaticlabs/prysm/v3/testing/endtoend/params"
 	"github.com/prysmaticlabs/prysm/v3/testing/endtoend/types"
 	"github.com/prysmaticlabs/prysm/v3/testing/require"
@@ -52,7 +53,7 @@ func e2eMinimal(t *testing.T, cfgo ...types.E2EConfigOpt) *testRunner {
 		ev.ColdStateCheckpoint,
 		ev.AltairForkTransition,
 		ev.BellatrixForkTransition,
-		ev.APIMiddlewareVerifyIntegrity,
+		evaluators.APIMiddlewareVerifyIntegrity,
 		ev.APIGatewayV1Alpha1VerifyIntegrity,
 		ev.FinishedSyncing,
 		ev.AllNodesHaveSameHead,
@@ -123,7 +124,7 @@ func e2eMainnet(t *testing.T, usePrysmSh, useMultiClient bool, cfgo ...types.E2E
 		ev.ColdStateCheckpoint,
 		ev.AltairForkTransition,
 		ev.BellatrixForkTransition,
-		ev.APIMiddlewareVerifyIntegrity,
+		evaluators.APIMiddlewareVerifyIntegrity,
 		ev.APIGatewayV1Alpha1VerifyIntegrity,
 		ev.FinishedSyncing,
 		ev.AllNodesHaveSameHead,
@@ -170,7 +171,7 @@ func scenarioEvals() []types.Evaluator {
 		ev.ColdStateCheckpoint,
 		ev.AltairForkTransition,
 		ev.BellatrixForkTransition,
-		ev.APIMiddlewareVerifyIntegrity,
+		evaluators.APIMiddlewareVerifyIntegrity,
 		ev.APIGatewayV1Alpha1VerifyIntegrity,
 		ev.FinishedSyncing,
 		ev.AllNodesHaveSameHead,
@@ -190,7 +191,7 @@ func scenarioEvalsMulti() []types.Evaluator {
 		ev.ColdStateCheckpoint,
 		ev.AltairForkTransition,
 		ev.BellatrixForkTransition,
-		ev.APIMiddlewareVerifyIntegrity,
+		evaluators.APIMiddlewareVerifyIntegrity,
 		ev.APIGatewayV1Alpha1VerifyIntegrity,
 		ev.FinishedSyncing,
 		ev.AllNodesHaveSameHead,
