+++ testing/endtoend/evaluators/beacon-api/util.go
@@ -1,0 +1,42 @@
+package evaluators
+
+import (
+	"bytes"
+	"encoding/json"
+	"fmt"
+	"net/http"
+
+	"github.com/prysmaticlabs/prysm/v3/testing/endtoend/params"
+)
+
+func doMiddlewareJSONGetRequest(template string, requestPath string, beaconNodeIdx int, dst interface{}) error {
+	basePath := fmt.Sprintf(template, params.TestParams.Ports.PrysmBeaconNodeGatewayPort+beaconNodeIdx)
+	httpResp, err := http.Get(
+		basePath + requestPath,
+	)
+	if err != nil {
+		return err
+	}
+	return json.NewDecoder(httpResp.Body).Decode(&dst)
+}
+
+func doMiddlewareJSONPostRequestV1(requestPath string, beaconNodeIdx int, postData, dst interface{}) error {
+	b, err := json.Marshal(postData)
+	if err != nil {
+		return err
+	}
+	basePath := fmt.Sprintf(v1MiddlewarePathTemplate, params.TestParams.Ports.PrysmBeaconNodeGatewayPort+beaconNodeIdx)
+	httpResp, err := http.Post(
+		basePath+requestPath,
+		"application/json",
+		bytes.NewBuffer(b),
+	)
+	if err != nil {
+		return err
+	}
+	return json.NewDecoder(httpResp.Body).Decode(&dst)
+}
+
+func buildFieldError(field, expected, actual string) error {
+	return fmt.Errorf("value of '%s' was expected to be '%s' but was '%s'", field, expected, actual)
+}
